+++ src/modules/mo/mo_dhl/Controller/Admin/OrderDHLController.php
@@ -54,6 +54,7 @@ class OrderDHLController extends \OxidEsales\Eshop\Application\Controller\Admin\
         $this->addTplParam('processes', Process::getAvailableProcesses());
         $this->addTplParam('ekp', $this->getEkp());
         $this->addTplParam('participationNumber', $this->getParticipationNumber());
+        $this->addTplParam('operator', $this->getOperator());
         $this->addTplParam('process', $this->getProcess());
         $this->addTplParam('remarks', $this->getRemarks());
         $this->addTplParam('labels', $this->getOrder()->moDHLGetLabels());
@@ -206,11 +207,24 @@ class OrderDHLController extends \OxidEsales\Eshop\Application\Controller\Admin\
     }
 
     /**
+     * @return string
+     */
+    protected function getOperator()
+    {
+        return (string)$this->getOrder()->oxorder__mo_dhl_operator->rawValue;
+    }
+
+    /**
      */
     public function save()
     {
         $db = \OxidEsales\Eshop\Core\DatabaseProvider::getDb(\OxidEsales\Eshop\Core\DatabaseProvider::FETCH_MODE_ASSOC);
-        $information = ['MO_DHL_EKP' => $this->validateEkp(), 'MO_DHL_PROCESS' => $this->validateProcessIdentifier(), 'MO_DHL_PARTICIPATION' => $this->validateParticipationNumber()];
+        $information = [
+            'MO_DHL_EKP' => $this->validateEkp(),
+            'MO_DHL_PROCESS' => $this->validateProcessIdentifier(),
+            'MO_DHL_PARTICIPATION' => $this->validateParticipationNumber(),
+            'MO_DHL_OPERATOR' => Registry::get(\OxidEsales\Eshop\Core\Request::class)->getRequestParameter('operator'),
+        ];
         $tuples = [];
         foreach ($information as $column => $value) {
             if (empty($value)) {